+++ Test/CoreSDK.Test/Shared/Extensibility/Implementation/Tracing/PortalDiagnosticsSenderTest.cs
@@ -94,6 +94,7 @@
             var trace = this.sendItems[0] as TraceTelemetry;
             Assert.IsNotNull(trace);
             Assert.AreEqual(diagnosticsInstrumentationKey, trace.Context.InstrumentationKey);
+            Assert.AreEqual("SDKTelemetry", trace.Context.Operation.SyntheticSource);
         }
 
         [TestMethod]