+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog will be used to generate documentation on [release notes page](ht
 Telemetry items that were filtered out by sampling are now indicated in debug output. Custom telemetry processors can now invoke
 method ```WriteTelemetry``` on ```TelemetryDebugWriter``` with ```filteredBy``` parameter to indicate in debug output that an
 item is being filtered out.
+- Internal Application Insights SDK traces will be marked as synthetic and have `SyntheticSource` equals to 'SDKTelemetry'
 
 ## Version 2.0.0-beta4