+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This changelog will be used to generate documentation on [release notes page](ht
 Telemetry items that were filtered out by sampling are now indicated in debug output. Custom telemetry processors can now invoke
 method ```WriteTelemetry``` on ```TelemetryDebugWriter``` with ```filteredBy``` parameter to indicate in debug output that an
 item is being filtered out.
+- DependencyTelemetry.Async property was removed.
+- DependencyTelemetry.Count property was removed.
+- When configuration is loaded from ApplicationInsights.config incorrect and broken elements are skiped. That includes both high level evelemts like TelemetryInitializers as well as individual properties.  
 - Internal Application Insights SDK traces will be marked as synthetic and have `SyntheticSource` equals to 'SDKTelemetry'
 
 ## Version 2.0.0-beta4